requirements.txt
crewai==0.51.1
crewai-tools==0.8.3
pyfiglet==1.0.2
langchain-chroma==0.1.3
langchain-ollama==0.1.1
python-dotenv==1.0.1
